Rename Exercise03.java to Exam Average
Write a program that prompts (using a Scanner to parse input) for the homework average, exam average, and final exam score for a student in this class. Your program should then print the final average and letter grade the student will receive. Be sure to follow all of the rules of the syllabus and validate your input (prompting the user again if needed). Include a text file (testcases.txt) in your project with the valid test cases, with expected results, you used. Of course, you should also test invalid cases, but you only have to report the valid ones.