Add files via upload
Write a program that prompts (using a Scanner for parsing input) for the homework average and exam average for a student in this class. Your program should then print a nicely formatted table of the final exam scores needed to receive each possible letter grade. Make sure that your output makes sense and only possible outcomes are displayed (e.g. it is not possible to get a negative score on the final). Be sure to follow all of the rules of the syllabus and validate your input (prompting the user again if needed). Include a text file (testcases.txt) in your project with the valid test cases, with expected results, you used. Of course, you should also test invalid cases, but you only have to report the valid ones.