Add files via upload
Write a program that allows the user to play Rock-paper-scissors-lizard-Spock (en.wikipedia.org/wiki/Rock-paper-scissors-lizard-Spock)
against the computer. The user should be able to play as many games as they wish without restarting your program. The program should
keep track of the score for the current session of the game and the overall total score for all sessions played. Think about how you
can efficiently represent the rules for winning.